+++ client/pages/mypage.js
@@ -2,7 +2,7 @@ import axios from 'axios';
 import React, { useEffect, useState } from 'react';
 import Image from 'next/image';
 import pic from '../public/honeycomb.png';
-import styles from '../styles/Tumbnail.module.css';
+import styles from '../styles/Post.module.css';
 import Link from 'next/link';
 
 export default function MyPage({ myPost, myScrap, alert, userInfo }) {
@@ -10,26 +10,6 @@ export default function MyPage({ myPost, myScrap, alert, userInfo }) {
     console.log(myScrap);
     console.log(alert);
     const [editBtn, setEditBtn] = useState(false);
-    // function getMyPage() {
-    //     axios
-    //         .get(`${process.env.NEXT_PUBLIC_URL}/myPage`, {
-    //             headers: { cookie: { accessToken: userInfo.accessToken }, 'Content-Type': 'application/json' },
-    //             withCredentials: true,
-    //         })
-    //         .then((res) => {
-    //             console.log(res.data.data);
-    //             setMyPost(res.data.data.myPost);
-    //             setMyScrap(res.data.data.myScrap);
-    //             setAlert(res.data.data.alert);
-    //         })
-    //         .catch((err) => {
-    //             return console.log('오류입니다!', err);
-    //         });
-    // }
-
-    // useEffect(() => {
-    //     getMyPage();
-    // }, []);
 
     function editMyPage() {
         axios.patch(`${process.env.NEXT_PUBLIC_URL}/mypage`, { userInfo }).then((res) => {
@@ -259,11 +239,10 @@ export default function MyPage({ myPost, myScrap, alert, userInfo }) {
 }
 
 export async function getServerSideProps(context) {
-    const cookies = context.req.headers.cookie;
-    console.log(cookies);
+    const token = context.req.headers.cookie;
     const apiUrl = `${process.env.NEXT_PUBLIC_URL}/mypage`;
     const res = await axios.get(apiUrl, {
-        headers: { cookie: cookies, 'Content-Type': 'application/json' },
+        headers: { cookie: token, 'Content-Type': 'application/json' },
     });
     const post = res.data.data.myPost;
     const scrap = res.data.data.myScrap;
