+++ client/src/components/Category.js
@@ -23,24 +23,20 @@ export default function Category({ setPost, init }) {
     return (
         <>
             {isClick ? (
-                <nav className="nav_area">
-                    <div className="nav_container">
-                        <button className="nav_btn" onClick={clickHandler}>
-                            <h1>카테고리</h1>
-                            <section>
-                                {categories.map((cate) => {
-                                    return (
-                                        <div onClick={(e) => categoryHandler(e)} key={cate}>
-                                            <button className="nav_items" onClick={(e) => categoryHandler(e)}>
-                                                <div className="category">{cate}</div>
-                                            </button>
-                                        </div>
-                                    );
-                                })}
-                            </section>
-                        </button>
-                    </div>
-                </nav>
+                <button className="nav_btn" onClick={clickHandler}>
+                    <h1>카테고리</h1>
+                    <section>
+                        {categories.map((cate) => {
+                            return (
+                                <div onClick={(e) => categoryHandler(e)} key={cate}>
+                                    <button className="nav_items" onClick={(e) => categoryHandler(e)}>
+                                        <div className="category">{cate}</div>
+                                    </button>
+                                </div>
+                            );
+                        })}
+                    </section>
+                </button>
             ) : (
                 <nav className="nav_area">
                     <div className="nav_container">
