+++ client/styles/Search.module.css
@@ -1,65 +1,111 @@
 .search_icon {
-    position: relative;
-    width: 3vw;
-    height: 3vw;
-}
-.search_box {
-    padding: 0vw;
+    left: 0;
+    top: 0;
     position: absolute;
-    left: 75vw;
-    top: 1vw;
-    width: 4vw;
-    height: 3vw;
+    width: 40px;
+    height: 40px;
+}
+
+.search_box {
+    padding: 0;
+    position: relative;
+    top: 20px;
+    left: 620px;
+    width: 40px;
+    height: 40px;
     transform: translate(-100%, -50%);
-    border-radius: 10vw;
+    border-radius: 20px;
     transition: 0.4s;
 }
 
 .search_box:hover {
-    box-shadow: 0px 0px 0.05vw 0.1vw rgba(244, 193, 65, 0.8);
-    width: 25vw;
+    box-shadow: 0px 0px 1px 3px rgba(244, 193, 65, 0.8);
+    width: 320px;
 }
 .search_txt {
     display: absolute;
-    padding: 0vw;
+    padding: 0;
     width: 0px;
     border: none;
     background: none;
     outline: none;
-    float: left;
-    margin-left: 1vw;
-    font-size: 1.5vw;
-    line-height: 3vw;
+    float: right;
+    margin-left: 10px;
+    font-size: 25px;
+    line-height: 45px;
     transition: 0.4s;
 }
 .search_box:hover > .search_txt {
-    width: 20vw;
-    padding: 0 0.6vw;
+    width: 280px;
+    padding: 0px 10px;
 }
 
 .search_container {
     position: relative;
     background-color: white;
 }
-.search_input {
-    position: absolute;
-    margin: -5.5vw 30vw;
-    width: 20vw;
-    height: 3vw;
-}
-
-@media screen and (max-width: 768px) {
+@media screen and (max-width: 1440px) {
     .search_icon {
+        left: 0;
+        top: 0;
+        position: absolute;
         width: 3vw;
         height: 3vw;
     }
     .search_box {
         padding: 0vw;
-        position: absolute;
-        left: 58vw;
+        position: relative;
+        left: 43vw;
         top: 1.5vw;
-        width: 4vw;
+        width: 3vw;
         height: 3vw;
+        transform: translate(-100%, -50%);
+        border-radius: 10vw;
+        transition: 0.4s;
+    }
+
+    .search_box:hover {
+        box-shadow: 0px 0px 0.05vw 0.1vw rgba(244, 193, 65, 0.8);
+        width: 25vw;
+    }
+    .search_txt {
+        display: absolute;
+        padding: 0vw;
+        width: 0px;
+        border: none;
+        background: none;
+        outline: none;
+        float: right;
+        font-size: 1.5vw;
+        line-height: 3vw;
+        transition: 0.4s;
+    }
+    .search_box:hover > .search_txt {
+        width: 21vw;
+        padding: 0 0.6vw;
+    }
+
+    .search_container {
+        position: relative;
+        background-color: white;
+    }
+}
+
+@media screen and (max-width: 768px) {
+    .search_icon {
+        left: 0;
+        top: 0;
+        position: absolute;
+        width: 4vw;
+        height: 4vw;
+    }
+    .search_box {
+        padding: 0vw;
+        position: absolute;
+        left: 56vw;
+        top: 2vw;
+        width: 4vw;
+        height: 4vw;
         transform: translate(-100%, -50%);
         border-radius: 10vw;
         transition: 0.4s;
@@ -75,14 +121,14 @@
         border: none;
         background: none;
         outline: none;
-        float: left;
+        float: right;
         margin-left: 1vw;
         font-size: 1.5vw;
-        line-height: 3vw;
+        line-height: 4vw;
         transition: 0.4s;
     }
     .search_box:hover > .search_txt {
-        width: 20vw;
+        width: 21vw;
         padding: 0 0.6vw;
     }
 
@@ -90,25 +136,20 @@
         position: relative;
         background-color: white;
     }
-    .search_input {
-        position: absolute;
-        margin: -5vw 30vw;
-        width: 20vw;
-        height: 2vw;
-    }
 }
 @media screen and (max-width: 425px) {
     .search_icon {
-        z-index: 0;
+        z-index: 10;
+        left: 0;
+        top: 0;
         position: absolute;
         width: 5vw;
         height: 5vw;
-        top: 0vw;
     }
     .search_box {
         padding: 1vw;
         position: absolute;
-        left: 45vw;
+        left: 55vw;
         top: 2vw;
         width: 2vw;
         height: 5vw;
@@ -127,7 +168,7 @@
         border: none;
         background: none;
         outline: none;
-        float: left;
+        float: right;
         margin-left: 1vw;
         font-size: 1vw;
         line-height: 3vw;
