+++ client/styles/Select.module.css
@@ -1,25 +1,40 @@
 .select_container {
     position: relative;
-    left: 70vw;
+    left: 850px;
+    top: 0;
 }
 .select {
-    width: 10vw;
-    height: 2vw;
+    width: 200px;
+    height: 40px;
 }
-@media screen and (max-width: 768px) {
+@media screen and (max-width: 1440px) {
     .select_container {
         position: relative;
-        left: 50vw;
+        left: 60vw;
+        top: 0vw;
     }
     .select {
         width: 15vw;
         height: 3vw;
     }
 }
+
+@media screen and (max-width: 768px) {
+    .select_container {
+        position: relative;
+        top: 0;
+        left: 48vw;
+    }
+    .select {
+        font-size: 1vw;
+        width: 14vw;
+        height: 4vw;
+    }
+}
 @media screen and (max-width: 425px) {
     .select_container {
         position: relative;
-        left: 46vw;
+        left: 45vw;
         top: -0.5vw;
     }
     .select {
